Move createAuthJsonSecret() around
It is used by both peer pods and the daemonset deployment mode.
Move it to `utils.go` for clarity. No functional change.

Reported-by: Patrik Fodor <patrik.fodor@ibm.com>
